# [Frontend] Don't log duplicate error stacktrace for every request in the batch