## Don't log duplicate error stacktrace for every request in the batch 

**EDIT:** Discussed with @joerunde, and we changed the solution. On the `client.py` if the engine is errored, then the  server should shut down, to do that we have to make sure to send a MQEngineDeadError despite the type of the exception. Previously, the server only shut down if the exception is a RuntimeError, and for those cases it does not log replicated stacktrace. With that we removed the exception MQEngineBatchError, because does not make sense anymore and we achieved the goal of this PR, because on launcher there is already an exception handler that does not replicate logs when it catches MQEngineBatchError.

---

Currently if there is an error in the engine when processing the current batch, the whole stacktrace ends up getting logged for every request. This PR address this issue to improve the server log readability.

## Steps to reproduce

I created a script that will keep sending requests to the server in parallel to make the engine busy and batching multiple requests.

``` python
import requests 
import json
from multiprocessing import Pool

prompts = ["how to make cheesecake", "How to make pizza", "Who is afraid of the big bad wolf?", "Who is the president of Brazil", "What is a capital of Spain", "Who is the president of USA"]


def do_generate(idx):
    
    while True:
        
        data = {
            'model': "ibm/merlinite-7b",
            'prompt': [prompts[idx]],
            'max_tokens': 200,
            "temperature": 0,
            "stream": False
        }

        res = requests.post('http://localhost:8000/v1/completions', data=json.dumps(data), headers={"Content-Type": "application/json"})

        print(json.dumps(res.json(), indent=2) )

if __name__ == "__main__": 
    indices = [1, 2, 3, 4, 5, 6]
    with Pool(len(indices)) as p:
        p.map(do_generate, indices)
    

```

On the vLLM side I added a hardcoded check to force an exception (not sure if there is a better way to do that). Basically, if it receives a request with the `max_tokens==123` then it raises an exception.

```python
@torch.inference_mode()
    @dump_input_when_exception(exclude_args=[0], exclude_kwargs=["self"])
    def execute_model(
        self,
        model_input: ModelInputForGPUWithSamplingMetadata,
        kv_caches: List[torch.Tensor],
        intermediate_tensors: Optional[IntermediateTensors] = None,
        num_steps: int = 1,
    ) -> Optional[Union[List[SamplerOutput], IntermediateTensors]]:
        
        for g in model_input.sampling_metadata.seq_groups:
            if g.sampling_params.max_tokens == 123: 
                raise Exception("FORCED EXCEPTION")
```

Then, just send a _poisoned_ curl with `max_tokens==123` to make the server crash.

```sh
curl http://localhost:8000/v1/completions -H "Content-Type: application/json"   -d '{
    "model": "ibm/merlinite-7b",
    "prompt": ["How to make pizza"],
    "max_tokens": 123,
    "temperature": 0
  }'

```

Following the log of the server for this scenario:

<details>

<summary>Server log</summary>

```log
INFO 10-01 18:43:01 api_server.py:520] vLLM API server version 0.6.1.post2
INFO 10-01 18:43:01 api_server.py:521] args: Namespace(host=None, port=8000, uvicorn_log_level='info', allow_credentials=False, allowed_origins=['*'], allowed_methods=['*'], allowed_headers=['*'], api_key=None, lora_modules=None, prompt_adapters=None, chat_template=None, response_role='assistant', ssl_keyfile=None, ssl_certfile=None, ssl_ca_certs=None, ssl_cert_reqs=0, root_path=None, middleware=[], return_tokens_as_token_ids=False, disable_frontend_multiprocessing=False, enable_auto_tool_choice=False, tool_call_parser=None, model='/shared_model_storage/transformers_cache/models--ibm--merlinite-7b/snapshots/233d12759d5bb9344231dafdb51310ec19d79c0e/', tokenizer=None, skip_tokenizer_init=False, revision=None, code_revision=None, tokenizer_revision=None, tokenizer_mode='auto', trust_remote_code=False, download_dir=None, load_format='auto', config_format='auto', dtype='auto', kv_cache_dtype='auto', quantization_param_path=None, max_model_len=None, guided_decoding_backend='outlines', distributed_executor_backend=None, worker_use_ray=False, pipeline_parallel_size=1, tensor_parallel_size=1, max_parallel_loading_workers=None, ray_workers_use_nsight=False, block_size=16, enable_prefix_caching=False, disable_sliding_window=False, use_v2_block_manager=False, num_lookahead_slots=0, seed=0, swap_space=4, cpu_offload_gb=0, gpu_memory_utilization=0.9, num_gpu_blocks_override=None, max_num_batched_tokens=None, max_num_seqs=256, max_logprobs=20, disable_log_stats=False, quantization=None, rope_scaling=None, rope_theta=None, enforce_eager=True, max_context_len_to_capture=None, max_seq_len_to_capture=8192, disable_custom_all_reduce=False, tokenizer_pool_size=0, tokenizer_pool_type='ray', tokenizer_pool_extra_config=None, limit_mm_per_prompt=None, enable_lora=False, max_loras=1, max_lora_rank=16, lora_extra_vocab_size=256, lora_dtype='auto', long_lora_scaling_factors=None, max_cpu_loras=None, fully_sharded_loras=False, enable_prompt_adapter=False, max_prompt_adapters=1, max_prompt_adapter_token=0, device='auto', num_scheduler_steps=1, scheduler_delay_factor=0.0, enable_chunked_prefill=None, speculative_model=None, speculative_model_quantization=None, num_speculative_tokens=None, speculative_draft_tensor_parallel_size=None, speculative_max_model_len=None, speculative_disable_by_batch_size=None, ngram_prompt_lookup_max=None, ngram_prompt_lookup_min=None, spec_decoding_acceptance_method='rejection_sampler', typical_acceptance_sampler_posterior_threshold=None, typical_acceptance_sampler_posterior_alpha=None, disable_logprobs_during_spec_decoding=None, model_loader_extra_config=None, ignore_patterns=[], preemption_mode=None, served_model_name=None, qlora_adapter_name_or_path=None, otlp_traces_endpoint=None, collect_detailed_traces=None, disable_async_output_proc=False, override_neuron_config=None, disable_log_requests=False, max_log_len=None, disable_fastapi_docs=False)
[...]
INFO 10-01 18:43:27 logger.py:36] Received request cmpl-c52dbd9a95114d1cbfa182e588138c86-0: prompt: 'How to make pizza', params: SamplingParams(n=1, best_of=1, presence_penalty=0.0, frequency_penalty=0.0, repetition_penalty=1.0, temperature=0.0, top_p=1.0, top_k=-1, min_p=0.0, seed=None, use_beam_search=False, length_penalty=1.0, early_stopping=False, stop=[], stop_token_ids=[], include_stop_str_in_output=False, ignore_eos=False, max_tokens=200, min_tokens=0, logprobs=None, prompt_logprobs=None, skip_special_tokens=True, spaces_between_special_tokens=True, truncate_prompt_tokens=None), prompt_token_ids: [1602, 298, 1038, 20727], lora_request: None, prompt_adapter_request: None.
INFO 10-01 18:43:27 engine.py:255] Added request cmpl-c52dbd9a95114d1cbfa182e588138c86-0.
INFO 10-01 18:43:27 logger.py:36] Received request cmpl-cf480cd01b3f45f9a44ee90f910b6dc7-0: prompt: 'Who is afraid of the big bad wolf?', params: SamplingParams(n=1, best_of=1, presence_penalty=0.0, frequency_penalty=0.0, repetition_penalty=1.0, temperature=0.0, top_p=1.0, top_k=-1, min_p=0.0, seed=None, use_beam_search=False, length_penalty=1.0, early_stopping=False, stop=[], stop_token_ids=[], include_stop_str_in_output=False, ignore_eos=False, max_tokens=200, min_tokens=0, logprobs=None, prompt_logprobs=None, skip_special_tokens=True, spaces_between_special_tokens=True, truncate_prompt_tokens=None), prompt_token_ids: [6526, 349, 8526, 302, 272, 2032, 2607, 24100, 28804], lora_request: None, prompt_adapter_request: None.
INFO 10-01 18:43:27 logger.py:36] Received request cmpl-a81e48bee0294b0b84cb89766e3747b1-0: prompt: 'Who is the president of Brazil', params: SamplingParams(n=1, best_of=1, presence_penalty=0.0, frequency_penalty=0.0, repetition_penalty=1.0, temperature=0.0, top_p=1.0, top_k=-1, min_p=0.0, seed=None, use_beam_search=False, length_penalty=1.0, early_stopping=False, stop=[], stop_token_ids=[], include_stop_str_in_output=False, ignore_eos=False, max_tokens=200, min_tokens=0, logprobs=None, prompt_logprobs=None, skip_special_tokens=True, spaces_between_special_tokens=True, truncate_prompt_tokens=None), prompt_token_ids: [6526, 349, 272, 4951, 302, 13250], lora_request: None, prompt_adapter_request: None.
INFO 10-01 18:43:27 logger.py:36] Received request cmpl-91fe755b3d0248ab94b2671eae5be62b-0: prompt: 'What is a capital of Spain', params: SamplingParams(n=1, best_of=1, presence_penalty=0.0, frequency_penalty=0.0, repetition_penalty=1.0, temperature=0.0, top_p=1.0, top_k=-1, min_p=0.0, seed=None, use_beam_search=False, length_penalty=1.0, early_stopping=False, stop=[], stop_token_ids=[], include_stop_str_in_output=False, ignore_eos=False, max_tokens=200, min_tokens=0, logprobs=None, prompt_logprobs=None, skip_special_tokens=True, spaces_between_special_tokens=True, truncate_prompt_tokens=None), prompt_token_ids: [1824, 349, 264, 5565, 302, 12567], lora_request: None, prompt_adapter_request: None.
INFO 10-01 18:43:27 logger.py:36] Received request cmpl-62a2bd68b2514c3997280fd68ae3bf38-0: prompt: 'Who is the president of us', params: SamplingParams(n=1, best_of=1, presence_penalty=0.0, frequency_penalty=0.0, repetition_penalty=1.0, temperature=0.0, top_p=1.0, top_k=-1, min_p=0.0, seed=None, use_beam_search=False, length_penalty=1.0, early_stopping=False, stop=[], stop_token_ids=[], include_stop_str_in_output=False, ignore_eos=False, max_tokens=200, min_tokens=0, logprobs=None, prompt_logprobs=None, skip_special_tokens=True, spaces_between_special_tokens=True, truncate_prompt_tokens=None), prompt_token_ids: [6526, 349, 272, 4951, 302, 592], lora_request: None, prompt_adapter_request: None.
INFO 10-01 18:43:27 metrics.py:351] Avg prompt throughput: 0.4 tokens/s, Avg generation throughput: 0.1 tokens/s, Running: 1 reqs, Swapped: 0 reqs, Pending: 0 reqs, GPU KV cache usage: 0.0%, CPU KV cache usage: 0.0%.
INFO 10-01 18:43:27 engine.py:255] Added request cmpl-cf480cd01b3f45f9a44ee90f910b6dc7-0.
INFO 10-01 18:43:27 engine.py:255] Added request cmpl-a81e48bee0294b0b84cb89766e3747b1-0.
INFO 10-01 18:43:27 engine.py:255] Added request cmpl-91fe755b3d0248ab94b2671eae5be62b-0.
INFO 10-01 18:43:27 engine.py:255] Added request cmpl-62a2bd68b2514c3997280fd68ae3bf38-0.
INFO:     ::1:51364 - "POST /v1/completions HTTP/1.1" 200 OK
INFO 10-01 18:43:28 logger.py:36] Received request cmpl-c5044c83c77b430d9038ecd2b7d1089e-0: prompt: 'Who is the president of us', params: SamplingParams(n=1, best_of=1, presence_penalty=0.0, frequency_penalty=0.0, repetition_penalty=1.0, temperature=0.0, top_p=1.0, top_k=-1, min_p=0.0, seed=None, use_beam_search=False, length_penalty=1.0, early_stopping=False, stop=[], stop_token_ids=[], include_stop_str_in_output=False, ignore_eos=False, max_tokens=200, min_tokens=0, logprobs=None, prompt_logprobs=None, skip_special_tokens=True, spaces_between_special_tokens=True, truncate_prompt_tokens=None), prompt_token_ids: [6526, 349, 272, 4951, 302, 592], lora_request: None, prompt_adapter_request: None.
INFO 10-01 18:43:28 engine.py:255] Added request cmpl-c5044c83c77b430d9038ecd2b7d1089e-0.
INFO:     ::1:51372 - "POST /v1/completions HTTP/1.1" 200 OK
INFO 10-01 18:43:28 logger.py:36] Received request cmpl-3af6a00b38db4987a36242d859e5cf93-0: prompt: 'Who is the president of us', params: SamplingParams(n=1, best_of=1, presence_penalty=0.0, frequency_penalty=0.0, repetition_penalty=1.0, temperature=0.0, top_p=1.0, top_k=-1, min_p=0.0, seed=None, use_beam_search=False, length_penalty=1.0, early_stopping=False, stop=[], stop_token_ids=[], include_stop_str_in_output=False, ignore_eos=False, max_tokens=200, min_tokens=0, logprobs=None, prompt_logprobs=None, skip_special_tokens=True, spaces_between_special_tokens=True, truncate_prompt_tokens=None), prompt_token_ids: [6526, 349, 272, 4951, 302, 592], lora_request: None, prompt_adapter_request: None.
INFO 10-01 18:43:28 engine.py:255] Added request cmpl-3af6a00b38db4987a36242d859e5cf93-0.
INFO:     ::1:51374 - "POST /v1/completions HTTP/1.1" 200 OK
INFO 10-01 18:43:29 logger.py:36] Received request cmpl-c9c4172a7de346d7ac1d38b3c2ef8446-0: prompt: 'Who is the president of us', params: SamplingParams(n=1, best_of=1, presence_penalty=0.0, frequency_penalty=0.0, repetition_penalty=1.0, temperature=0.0, top_p=1.0, top_k=-1, min_p=0.0, seed=None, use_beam_search=False, length_penalty=1.0, early_stopping=False, stop=[], stop_token_ids=[], include_stop_str_in_output=False, ignore_eos=False, max_tokens=200, min_tokens=0, logprobs=None, prompt_logprobs=None, skip_special_tokens=True, spaces_between_special_tokens=True, truncate_prompt_tokens=None), prompt_token_ids: [6526, 349, 272, 4951, 302, 592], lora_request: None, prompt_adapter_request: None.
INFO 10-01 18:43:29 engine.py:255] Added request cmpl-c9c4172a7de346d7ac1d38b3c2ef8446-0.
INFO:     ::1:51380 - "POST /v1/completions HTTP/1.1" 200 OK
INFO 10-01 18:43:29 logger.py:36] Received request cmpl-dbc219febf5f4375ae74c0eed37c30a0-0: prompt: 'Who is the president of us', params: SamplingParams(n=1, best_of=1, presence_penalty=0.0, frequency_penalty=0.0, repetition_penalty=1.0, temperature=0.0, top_p=1.0, top_k=-1, min_p=0.0, seed=None, use_beam_search=False, length_penalty=1.0, early_stopping=False, stop=[], stop_token_ids=[], include_stop_str_in_output=False, ignore_eos=False, max_tokens=200, min_tokens=0, logprobs=None, prompt_logprobs=None, skip_special_tokens=True, spaces_between_special_tokens=True, truncate_prompt_tokens=None), prompt_token_ids: [6526, 349, 272, 4951, 302, 592], lora_request: None, prompt_adapter_request: None.
INFO 10-01 18:43:29 engine.py:255] Added request cmpl-dbc219febf5f4375ae74c0eed37c30a0-0.
INFO:     ::1:51388 - "POST /v1/completions HTTP/1.1" 200 OK
INFO 10-01 18:43:30 logger.py:36] Received request cmpl-97b0729b39ac464e98518e9773d42fff-0: prompt: 'Who is the president of us', params: SamplingParams(n=1, best_of=1, presence_penalty=0.0, frequency_penalty=0.0, repetition_penalty=1.0, temperature=0.0, top_p=1.0, top_k=-1, min_p=0.0, seed=None, use_beam_search=False, length_penalty=1.0, early_stopping=False, stop=[], stop_token_ids=[], include_stop_str_in_output=False, ignore_eos=False, max_tokens=200, min_tokens=0, logprobs=None, prompt_logprobs=None, skip_special_tokens=True, spaces_between_special_tokens=True, truncate_prompt_tokens=None), prompt_token_ids: [6526, 349, 272, 4951, 302, 592], lora_request: None, prompt_adapter_request: None.
INFO 10-01 18:43:30 engine.py:255] Added request cmpl-97b0729b39ac464e98518e9773d42fff-0.
INFO:     ::1:51402 - "POST /v1/completions HTTP/1.1" 200 OK
INFO 10-01 18:43:30 logger.py:36] Received request cmpl-21dd2f2709064410bcd6c66eda41d1da-0: prompt: 'Who is the president of us', params: SamplingParams(n=1, best_of=1, presence_penalty=0.0, frequency_penalty=0.0, repetition_penalty=1.0, temperature=0.0, top_p=1.0, top_k=-1, min_p=0.0, seed=None, use_beam_search=False, length_penalty=1.0, early_stopping=False, stop=[], stop_token_ids=[], include_stop_str_in_output=False, ignore_eos=False, max_tokens=200, min_tokens=0, logprobs=None, prompt_logprobs=None, skip_special_tokens=True, spaces_between_special_tokens=True, truncate_prompt_tokens=None), prompt_token_ids: [6526, 349, 272, 4951, 302, 592], lora_request: None, prompt_adapter_request: None.
INFO 10-01 18:43:30 engine.py:255] Added request cmpl-21dd2f2709064410bcd6c66eda41d1da-0.
INFO:     ::1:51412 - "POST /v1/completions HTTP/1.1" 200 OK
INFO 10-01 18:43:31 logger.py:36] Received request cmpl-6837e12c5c3c4ac5aaf6bbe816dd2a69-0: prompt: 'Who is the president of us', params: SamplingParams(n=1, best_of=1, presence_penalty=0.0, frequency_penalty=0.0, repetition_penalty=1.0, temperature=0.0, top_p=1.0, top_k=-1, min_p=0.0, seed=None, use_beam_search=False, length_penalty=1.0, early_stopping=False, stop=[], stop_token_ids=[], include_stop_str_in_output=False, ignore_eos=False, max_tokens=200, min_tokens=0, logprobs=None, prompt_logprobs=None, skip_special_tokens=True, spaces_between_special_tokens=True, truncate_prompt_tokens=None), prompt_token_ids: [6526, 349, 272, 4951, 302, 592], lora_request: None, prompt_adapter_request: None.
INFO 10-01 18:43:31 engine.py:255] Added request cmpl-6837e12c5c3c4ac5aaf6bbe816dd2a69-0.
INFO:     ::1:51426 - "POST /v1/completions HTTP/1.1" 200 OK
INFO 10-01 18:43:31 logger.py:36] Received request cmpl-69faaae7ba6b48e58655c07a2e7ba046-0: prompt: 'How to make pizza', params: SamplingParams(n=1, best_of=1, presence_penalty=0.0, frequency_penalty=0.0, repetition_penalty=1.0, temperature=0.0, top_p=1.0, top_k=-1, min_p=0.0, seed=None, use_beam_search=False, length_penalty=1.0, early_stopping=False, stop=[], stop_token_ids=[], include_stop_str_in_output=False, ignore_eos=False, max_tokens=123, min_tokens=0, logprobs=None, prompt_logprobs=None, skip_special_tokens=True, spaces_between_special_tokens=True, truncate_prompt_tokens=None), prompt_token_ids: [1602, 298, 1038, 20727], lora_request: None, prompt_adapter_request: None.
INFO 10-01 18:43:31 engine.py:255] Added request cmpl-69faaae7ba6b48e58655c07a2e7ba046-0.
INFO 10-01 18:43:31 model_runner_base.py:120] Writing input of failed execution to /tmp/err_execute_model_input_20241001-184331.pkl...
INFO 10-01 18:43:31 model_runner_base.py:141] Completed writing input of failed execution to /tmp/err_execute_model_input_20241001-184331.pkl.
INFO:     ::1:51348 - "POST /v1/completions HTTP/1.1" 500 Internal Server Error
ERROR 10-01 18:43:31 engine.py:130] Exception('Error in model execution (input dumped to /tmp/err_execute_model_input_20241001-184331.pkl): FORCED EXCEPTION')
ERROR 10-01 18:43:31 engine.py:130] Traceback (most recent call last):
ERROR 10-01 18:43:31 engine.py:130]   File "/tmp/vllm/vllm/worker/model_runner_base.py", line 116, in _wrapper
ERROR 10-01 18:43:31 engine.py:130]     return func(*args, **kwargs)
ERROR 10-01 18:43:31 engine.py:130]   File "/tmp/vllm/vllm/worker/model_runner.py", line 1548, in execute_model
ERROR 10-01 18:43:31 engine.py:130]     raise Exception("FORCED EXCEPTION")
ERROR 10-01 18:43:31 engine.py:130] Exception: FORCED EXCEPTION
ERROR 10-01 18:43:31 engine.py:130] 
ERROR 10-01 18:43:31 engine.py:130] The above exception was the direct cause of the following exception:
ERROR 10-01 18:43:31 engine.py:130] 
ERROR 10-01 18:43:31 engine.py:130] Traceback (most recent call last):
ERROR 10-01 18:43:31 engine.py:130]   File "/tmp/vllm/vllm/engine/multiprocessing/engine.py", line 128, in start
ERROR 10-01 18:43:31 engine.py:130]     self.run_engine_loop()
ERROR 10-01 18:43:31 engine.py:130]   File "/tmp/vllm/vllm/engine/multiprocessing/engine.py", line 188, in run_engine_loop
ERROR 10-01 18:43:31 engine.py:130]     request_outputs = self.engine_step()
ERROR 10-01 18:43:31 engine.py:130]   File "/tmp/vllm/vllm/engine/multiprocessing/engine.py", line 207, in engine_step
ERROR 10-01 18:43:31 engine.py:130]     raise e
ERROR 10-01 18:43:31 engine.py:130]   File "/tmp/vllm/vllm/engine/multiprocessing/engine.py", line 198, in engine_step
ERROR 10-01 18:43:31 engine.py:130]     return self.engine.step()
ERROR 10-01 18:43:31 engine.py:130]   File "/tmp/vllm/vllm/engine/llm_engine.py", line 1228, in step
ERROR 10-01 18:43:31 engine.py:130]     outputs = self.model_executor.execute_model(
ERROR 10-01 18:43:31 engine.py:130]   File "/tmp/vllm/vllm/executor/gpu_executor.py", line 130, in execute_model
ERROR 10-01 18:43:31 engine.py:130]     output = self.driver_worker.execute_model(execute_model_req)
ERROR 10-01 18:43:31 engine.py:130]   File "/tmp/vllm/vllm/worker/worker_base.py", line 327, in execute_model
ERROR 10-01 18:43:31 engine.py:130]     output = self.model_runner.execute_model(
…