@wallashss 2024-10-03
Thanks for the review @joerunde! See my comments and check if it make sense. I am totally open to review ideas and change the implementation.